Bug #437: Squid can't handle files larger than 2 GB.
Kill unneeded size_t -> ssize_t typecasts

+++ src/fs/ufs/store_io_ufs.c
@@ -221,7 +221,7 @@ storeUfsReadDone(int fd, const char *buf, int len, int errflag, void *my_data)
     sio->read.callback = NULL;
     sio->read.callback_data = NULL;
     if (cbdataValid(their_data))
-	callback(their_data, buf, (size_t) rlen);
+	callback(their_data, buf, rlen);
     cbdataUnlock(their_data);
 }
